seperate models out by device type
