+++ package.json
@@ -13,8 +13,8 @@
     "url": "https://github.com/donavanbecker/homebridge-meross/issues"
   },
   "engines": {
-    "node": ">=14.15.4",
-    "homebridge": ">=1.3.1"
+    "node": ">=14.16.0",
+    "homebridge": ">=1.3.2"
   },
   "main": "dist/index.js",
   "scripts": {
@@ -48,11 +48,9 @@
     "@typescript-eslint/eslint-plugin": "^4.16.1",
     "@typescript-eslint/parser": "^4.16.1",
     "eslint": "^7.21.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "homebridge": "^1.3.1",
+    "homebridge": "^1.3.2",
     "homebridge-config-ui-x": "^4.39.1",
     "nodemon": "^2.0.7",
-    "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
     "ts-node": "^9.1.1",
     "typescript": "^4.2.2",
